requirements.txt
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp_socks==0.11.0
aiosignal==1.4.0
asttokens==3.0.1
asyncio==4.0.0
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
csfloat_api==1.3.0
debugpy==1.8.19
decorator==5.2.1
executing==2.2.1
frozenlist==1.8.0
idna==3.11
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
matplotlib-inline==0.2.1
multidict==6.7.0
nest-asyncio==1.6.0
numpy==2.4.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.1
pure_eval==0.2.3
pycares==5.0.1
pycparser==2.23
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-socks==2.8.0
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
six==1.17.0
stack-data==0.6.3
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.2.14
yarl==1.22.0
